Changed impl structure
