+++ include/exl/mixed.hpp
@@ -12,7 +12,7 @@
 
 #include <exl/matchers.hpp>
 #include <exl/in_place.hpp>
-#include <exl/impl/mixed_storage_operations.hpp>
+#include <exl/impl/mixed/mixed_storage_operations.hpp>
 
 namespace exl
 {
