+++ tests/CMakeLists.txt
@@ -5,15 +5,17 @@
 
 add_executable(exl-test
         # Utilities
-        utils/ClassMock.cpp
+        test-utils/ClassMock.cpp
         # Tests
         tests.cpp
-        type_list.cpp
-        matchers.cpp
-        mixed.cpp
-        option.cpp
-        box.cpp
+        mixed/impl/type_list.cpp
+        matchers/matchers.cpp
+        mixed/mixed.cpp
+        option/option.cpp
+        box/box.cpp
 )
+
+include_directories(exl-test ${CMAKE_CURRENT_LIST_DIR}/test-utils)
 
 target_link_libraries(exl-test PRIVATE exl)
 target_link_libraries(exl-test PRIVATE Catch2)
@@ -41,12 +43,13 @@ add_test(NAME exl-test COMMAND exl-test)
 
 function(add_termination_test TEST_NAME TEST_FILE)
     add_executable(${TEST_NAME} ${TEST_FILE})
+    include_directories(${TEST_NAME} ${CMAKE_CURRENT_LIST_DIR}/test-utils)
     target_link_libraries(${TEST_NAME} PRIVATE exl)
     target_link_libraries(${TEST_NAME} PRIVATE exl-coverage-config)
     add_test(NAME ${TEST_NAME} COMMAND ${TEST_NAME})
     set_tests_properties(${TEST_NAME} PROPERTIES WILL_FAIL TRUE)
 endfunction()
 
-add_termination_test(exl-mixed-invalid-unwrap-test mixed_invalid_unwrap_test.cpp)
-add_termination_test(exl-mixed-invalid-unwrap-exact-test mixed_invalid_unwrap_exact_test.cpp)
-add_termination_test(exl-box-invalid-dereferencing-test box_invalid_dereferencing_test.cpp)
+add_termination_test(exl-mixed-invalid-unwrap-test mixed/mixed_invalid_unwrap_test.cpp)
+add_termination_test(exl-mixed-invalid-unwrap-exact-test mixed/mixed_invalid_unwrap_exact_test.cpp)
+add_termination_test(exl-box-invalid-dereferencing-test box/box_invalid_dereferencing_test.cpp)
