+++ tests/box/box_invalid_dereferencing_test.cpp
@@ -3,12 +3,18 @@
 // Distributed under the Boost Software License, Version 1.0.
 // (See accompanying file LICENSE or copy at https://www.boost.org/LICENSE_1_0.txt)
 
-#include <exl/mixed.hpp>
+#include <exl/box.hpp>
 
-#include "utils/termination_test.hpp"
+#include <termination_test.hpp>
+
+#include <AllocObject.hpp>
+
+using namespace exl::test;
 
 void termination_test()
 {
-    exl::mixed<int, std::runtime_error, std::exception> m(std::runtime_error("boom"));
-    m.unwrap_exact<std::exception>();
+    auto boxed = exl::box<AlwaysBadAllocObject>::make();
+    auto& unwrapped = boxed.get();
+
+    (void) unwrapped;
 }
