+++ tests/mixed/mixed_invalid_unwrap_exact_test.cpp
@@ -3,17 +3,12 @@
 // Distributed under the Boost Software License, Version 1.0.
 // (See accompanying file LICENSE or copy at https://www.boost.org/LICENSE_1_0.txt)
 
-#include <exl/box.hpp>
+#include <exl/mixed.hpp>
 
-#include "utils/termination_test.hpp"
-#include "utils/AllocObject.hpp"
-
-using namespace exl::test;
+#include <termination_test.hpp>
 
 void termination_test()
 {
-    auto boxed = exl::box<AlwaysBadAllocObject>::make();
-    auto& unwrapped = boxed.get();
-
-    (void) unwrapped;
+    exl::mixed<int, std::runtime_error, std::exception> m(std::runtime_error("boom"));
+    m.unwrap_exact<std::exception>();
 }
