+++ tests/test-utils/ClassMock.cpp
