+++ tests/test-utils/ClassMock.hpp
